Put all DBs on the same named network
I have made a change to expose the port on which the MySQL server is running to the host, and to
network all the containers together by putting them in the same named Docker network.

This feature is useful when there is a different Docker container which needs to connect to one of
the DB containers.